requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.1.0
gotrue>=1.3.0,<2.0.0
google-generativeai==0.3.2
httpx>=0.24,<0.26
PyPDF2==3.0.1
pdfplumber==0.11.6
python-docx==1.1.2
python-multipart==0.0.9
pydantic==2.6.3 